build java and tests in Docs subproj
