+++ cdm-test-utils/src/main/java/ucar/unidata/util/test/TestLogger.java
@@ -1,4 +1,4 @@
-package examples.tutorials;
+package ucar.unidata.util.test;
 
 import java.util.Stack;
